Make create fixtures local
